+++ src/interactive/scala/tools/nsc/interactive/Replayer.scala
@@ -1,4 +1,4 @@
-package scala.tools.nsc.io
+package scala.tools.nsc.interactive
 
 import java.io.{Reader, Writer}
